refactor validation test helpers
